Rename api file to more consistent description
